chore: use rem in css
